+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+markers = slow: marks tests as slow (deselect with '-m "not slow"')
+doctest_optionflags = NORMALIZE_WHITESPACE
+filterwarnings =
+    error::UserWarning
