+++ requirements.txt
@@ -1,3 +1,6 @@
 github3.py==1.3.0
 python-dotenv==0.15.0
 python-dateutil==2.8.1
+pytest==6.2.2
+pytest-cov==2.11.1
+pytest-html==3.1.1
